requirements.txt
torch==2.0.1
transformers==4.35.2
sentence-transformers==2.2.2
chromadb==0.4.18
pymupdf==1.23.5
numpy==1.24.3
pandas==2.0.3
qdrant-client==1.7.0
python-dotenv==1.0.0
pydantic==2.4.2
langchain==0.0.310
langchain-community==0.0.6
requests==2.31.0
tqdm==4.66.1
pymongo==4.6.1
fastapi==0.109.0
uvicorn==0.27.0
PyMuPDF>=1.23.0
paddleocr==3.3.0
pdf2image==1.16.3
Pillow==10.1.0
